Query the IGVF Catalog via ArangoDB

edges() and nodes() identify edges or nodes in the data base.

db_gene_variants() locates variants associated with a (Ensembl) gene identifier.

db_gene_elements() locates genomic elements associated with a (Ensembl) gene identifier.

Usage

db_edges(
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

db_nodes(
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

db_gene_variants(
  gene_id,
  threshold,
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

db_gene_elements(
  gene_id,
  threshold,
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

Arguments

username

character(1) ArangoDB user name. Default: "guest".

password

character(1) ArangoDB password. Default: "guestigvfcatalog". A better practice is to use an environment variable to record the password, rather than encoding in a script, so password = Sys.getenv("RIGVF_ARANGODB_PASSWORD").

gene_id

character(1) Ensembl gene identifier

threshold

numeric(1) minimum association statistic, minus log10 p-value for variants, and score for elements

Value

edges() and nodes() return a tibble with the edge or node name and count of occurrences in the database.

db_gene_variants() returns a tibble summarizing variants associated with the gene.

db_gene_elements() returns a tibble summarizing genomic elements associated with the gene.

Examples

db_edges()
#> # A tibble: 34 × 2
#>    name                                   count
#>    <chr>                                  <dbl>
#>  1 variants_variants                 5362104968
#>  2 variants_coding_variants           239116929
#>  3 coding_variants_proteins           238602987
#>  4 variants_genes_terms                95728528
#>  5 variants_genes                      95728517
#>  6 genomic_elements_genes              31514158
#>  7 transcripts_genes_structure          4835125
#>  8 genes_genes                          3553547
#>  9 variants_proteins                    2927999
#> 10 mm_transcripts_mm_genes_structure    2417458
#> # ℹ 24 more rows

db_nodes()
#> # A tibble: 19 × 2
#>    name                     count
#>    <chr>                    <dbl>
#>  1 variants            1185310425
#>  2 coding_variants      239116929
#>  3 mm_variants          101894574
#>  4 genomic_elements      17833721
#>  5 genes_structure        4835125
#>  6 mm_genes_structure     2417758
#>  7 mm_genomic_elements     926826
#>  8 ontology_terms          731807
#>  9 proteins                290309
#> 10 transcripts             274031
#> 11 mm_transcripts          149547
#> 12 genes                    69222
#> 13 mm_genes                 56941
#> 14 studies                  22690
#> 15 drugs                     4613
#> 16 pathways                  2711
#> 17 complexes                 1681
#> 18 motifs                     401
#> 19 donors                     231

db_gene_variants("ENSG00000106633", threshold = 4.0)
#> # A tibble: 1,000 × 14
#>    `_key`           `_id` `_from` `_to` `_rev` biological_context chr    p_value
#>    <chr>            <chr> <chr>   <chr> <chr>  <chr>              <chr>    <dbl>
#>  1 a6cbc153279e78b… vari… varian… gene… _i42_… esophagus muscula… chr7  5.82e-12
#>  2 fe37460fd352499… vari… varian… gene… _i42_… esophagus muscula… chr7  3.55e- 5
#>  3 541cb07e388227c… vari… varian… gene… _i42_… esophagus muscula… chr7  9.43e-12
#>  4 f926f23ed97a1ae… vari… varian… gene… _i42_… esophagus muscula… chr7  9.43e-12
#>  5 4600cd627888076… vari… varian… gene… _i42_… esophagus muscula… chr7  4.17e-12
#>  6 0642a7286da19b2… vari… varian… gene… _i42_… esophagus muscula… chr7  3.58e- 5
#>  7 b2b44a8a0e56c9b… vari… varian… gene… _i42_… esophagus muscula… chr7  2.94e- 5
#>  8 3808659e1f51104… vari… varian… gene… _i42_… esophagus muscula… chr7  2.68e- 5
#>  9 838b2fe5120da15… vari… varian… gene… _i42_… esophagus muscula… chr7  2.03e- 5
#> 10 a6cd41ca17f99aa… vari… varian… gene… _i42_… esophagus muscula… chr7  1.09e-12
#> # ℹ 990 more rows
#> # ℹ 6 more variables: log10pvalue <dbl>, effect_size <dbl>, pval_beta <dbl>,
#> #   label <chr>, source <chr>, source_url <chr>

db_gene_elements("ENSG00000106633", threshold = 0.5)
#> # A tibble: 103 × 12
#>    `_key`      `_id` `_from` `_to` `_rev` score source source_url file_accession
#>    <chr>       <chr> <chr>   <chr> <chr>  <dbl> <chr>  <chr>      <chr>         
#>  1 genic_chr7… geno… genomi… gene… _jI7z… 0.524 ENCOD… https://w… ENCFF009QHG   
#>  2 genic_chr7… geno… genomi… gene… _jI7z… 0.961 ENCOD… https://w… ENCFF009QHG   
#>  3 genic_chr7… geno… genomi… gene… _jI7z… 0.621 ENCOD… https://w… ENCFF009QHG   
#>  4 promoter_c… geno… genomi… gene… _jI7z… 0.974 ENCOD… https://w… ENCFF009QHG   
#>  5 genic_chr7… geno… genomi… gene… _jI70… 0.989 ENCOD… https://w… ENCFF057RTZ   
#>  6 promoter_c… geno… genomi… gene… _jI70… 0.869 ENCOD… https://w… ENCFF057RTZ   
#>  7 genic_chr7… geno… genomi… gene… _jI71… 0.942 ENCOD… https://w… ENCFF059DHV   
#>  8 genic_chr7… geno… genomi… gene… _jI71… 0.574 ENCOD… https://w… ENCFF059DHV   
#>  9 genic_chr7… geno… genomi… gene… _jI71… 0.828 ENCOD… https://w… ENCFF059DHV   
#> 10 promoter_c… geno… genomi… gene… _jI71… 0.929 ENCOD… https://w… ENCFF059DHV   
#> # ℹ 93 more rows
#> # ℹ 3 more variables: biological_context <chr>, name <chr>, inverse_name <chr>