Query the 'IGVF' 'ArangoDB'

edges() and nodes() identify edges or nodes in the data base.

db_gene_variants() locates variants associated with a (Ensembl) gene identifier.

Usage

db_edges(
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

db_nodes(
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

db_gene_variants(
  gene_id,
  threshold,
  username = rigvf_config$get("username"),
  password = rigvf_config$get("password")
)

Arguments

username

character(1) ArangoDB user name. Default: "guest".

password

character(1) ArangoDB password. Default: "guestigvfcatalog". A better practice is to use an environment variable to record the password, rather than encoding in a script, so password = Sys.getenv("RIGVF_ARANGODB_PASSWORD").

gene_id

character(1) Ensembl gene identifier.

threshold

numeric(1) minimum score associated with the variant.

Value

edges() and nodes() return a tibble with the edge or node name and count of occurrences in the database.

db_gene_variants() returns a tibble summarizing variants associated with the gene.

Examples

db_edges()
#> # A tibble: 36 × 2
#>    name                                            count
#>    <chr>                                           <dbl>
#>  1 variants_variants                          5979387121
#>  2 coding_variants_proteins                    239108968
#>  3 variants_coding_variants                    239108968
#>  4 variants_genes_terms                         95728528
#>  5 variants_genes                               95728528
#>  6 regulatory_regions_genes                     30197460
#>  7 regulatory_regions_genes_biosamples_donors   21075288
#>  8 regulatory_regions_genes_biosamples          19703741
#>  9 genes_genes                                   3553549
#> 10 variants_proteins                             2927144
#> # ℹ 26 more rows

db_nodes()
#> # A tibble: 17 × 2
#>    name                       count
#>    <chr>                      <dbl>
#>  1 variants              1304691412
#>  2 coding_variants        239107830
#>  3 mm_variants            101894574
#>  4 regulatory_regions      17138182
#>  5 mm_regulatory_regions    3162620
#>  6 ontology_terms            728868
#>  7 proteins                  294001
#>  8 transcripts               274031
#>  9 mm_transcripts            149547
#> 10 genes                      69222
#> 11 mm_genes                   56941
#> 12 studies                    22690
#> 13 drugs                       4613
#> 14 pathways                    2629
#> 15 complexes                   1527
#> 16 motifs                       401
#> 17 donors                       231

db_gene_variants("ENSG00000106633", 0.85)
#> # A tibble: 40 × 9
#>    `_key`              `_id` `_from` `_to` `_rev` `score:long` source source_url
#>    <chr>               <chr> <chr>   <chr> <chr>         <dbl> <chr>  <chr>     
#>  1 genic_chr7_4415452… regu… regula… gene… _g5CU…        0.989 ENCOD… https://w…
#>  2 promoter_chr7_4415… regu… regula… gene… _g5CU…        0.869 ENCOD… https://w…
#>  3 genic_chr7_4414584… regu… regula… gene… _g5CU…        0.948 ENCOD… https://w…
#>  4 promoter_chr7_4415… regu… regula… gene… _g5CU…        1.00  ENCOD… https://w…
#>  5 promoter_chr7_4415… regu… regula… gene… _g5CU…        1.00  ENCOD… https://w…
#>  6 intergenic_chr7_44… regu… regula… gene… _g5CU…        0.959 ENCOD… https://w…
#>  7 genic_chr7_4415544… regu… regula… gene… _g5CV…        0.942 ENCOD… https://w…
#>  8 promoter_chr7_4415… regu… regula… gene… _g5CV…        0.929 ENCOD… https://w…
#>  9 intergenic_chr7_44… regu… regula… gene… _g5CV…        0.936 ENCOD… https://w…
#> 10 promoter_chr7_4415… regu… regula… gene… _g5CW…        0.966 ENCOD… https://w…
#> # ℹ 30 more rows
#> # ℹ 1 more variable: biological_context <chr>